--- app/ns-ui-widgets-category/search-bar/basics/article.md
@@ -1,5 +1,0 @@
-The example shows how to create SerachBr component via XML and handle submit and clear events.
-
-<snippet id='search-bar-basics-xml'/>
-<snippet id='search-bar-basics-code'/>
-<snippet id='search-bar-basics-code-ts'/>
