--- app/ns-ui-widgets-category/search-bar/basics/basics-page.css
@@ -1,4 +1,0 @@
-
-button, label, stack-layout {
-    horizontal-align: center;    
-}
