--- app/ns-ui-widgets-category/search-bar/basics/basics-ts-page.xml
@@ -1,8 +1,0 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo" loaded="onPageLoaded">
-    <Page.actionBar>
-		<ActionBar title="Getting Started"/>
-	</Page.actionBar>
-		<StackLayout class="p-20">
-       <SearchBar id="searchBar" hint="{{sbHint}}" text="{{sbText}}" clear="onClear" submit="onSubmit" />
-    </StackLayout>
-</Page>
