+++ app/ns-ui-widgets-category/search-bar/end.md
@@ -1,3 +1,35 @@
+### Creating SearchBar component via Code-Behind
 
-**API Reference for the** [SearchBar Class](https://docs.nativescript.org/api-reference/classes/_ui_search_bar_.searchbar)
+<snippet id='creating-search-bar-xml'/>
+<snippet id='creating-search-bar-code'/>
+<snippet id='creating-search-bar-code-ts'/>
 
+## Properties
+
+| Name     | Type    | Description    |
+|----------|---------|----------------|
+| `hint`   | `string` | Gets or sets the text of the search bar text field hint/placeholder. |
+| `text`   | `string` | Gets or sets a search bar text. |
+| `textFieldBackgroundColor`   | `Color` | Gets or sets the TextField background color of the SearchBar component. |
+| `textFieldHintColor`   | `Color` | Gets or sets the TextField Hint color of the SearchBar component. |
+
+## Events
+
+| Name     | Description    |
+|----------|----------------|
+| `clear`  | Emitted when the search is cleared (via the clear button). |
+| `submit` | Emitted when the serach term is submitted.   |
+
+
+## API References
+
+| Name     | Type    | 
+|----------|---------|
+| [tns-core-modules/ui/search-bar](https://docs.nativescript.org/api-reference/modules/_ui_search_bar_) | `Module` | 
+| [SearchBar](https://docs.nativescript.org/api-reference/classes/_ui_search_bar_.searchbar) | `Class` | 
+
+## Native Component
+
+| Android               | iOS      |
+|:----------------------|:---------|
+| [android.widget.SearchView](https://developer.android.com/reference/android/widget/SearchView) | [UISearchBar](https://developer.apple.com/documentation/uikit/uisearchbar) |
