--- app/ns-ui-widgets-category/search-bar/events/article.md
@@ -1,6 +1,0 @@
-Example showing how to use SearchBar `submit` and `clear` events.
-
-
-<snippet id='clear-text-search-bar-html'/>
-<snippet id='clear-search-bar-submit'/>
-<snippet id='clear-search-bar-submit-ts'/>
