--- app/ns-ui-widgets-category/search-bar/events/events-page.xml
@@ -1,18 +1,0 @@
-<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo">
-    <Page.actionBar>
-		<ActionBar title="Events"/>
-	</Page.actionBar>
-    <GridLayout rows="auto *">
-        <!-- >> clear-text-search-bar-html -->
-        <SearchBar row="0" hint="Search for a country and press enter" clear="onClear" submit="onSubmit"/>
-
-		<ListView row="1" items="{{ myItems }}"  class="list-group">
-			<ListView.itemTemplate>
-				<GridLayout class="item list-group-item">
-                    <Label text="{{ name }}" class="list-group-item-heading"/>
-                </GridLayout>
-			</ListView.itemTemplate>
-		</ListView>
-        <!-- << clear-text-search-bar-html -->
-    </GridLayout>
-</Page>
