--- app/ns-ui-widgets-category/search-bar/events/events-ts-page.css
@@ -1,13 +1,0 @@
-
-button, label, stack-layout {
-    horizontal-align: center;    
-}
-
-
-
-GridLayout {
-    padding: 0;
-    margin-left: 0;
-    margin-top: 10;
-    margin-bottom: 10;
-}
