+++ app/ns-ui-widgets-category/search-bar/overview.md
@@ -1,5 +1,2 @@
-The SearchBar module represents a UI component similar to `UISearchBar` in iOS and `SearchView` for Android, both of which allow you to to create a simple filter for the the content in the app. 
-This module gives you the simple way to handle a `submit` event and a `clear` event.
-
-<snippet id='search-bar-require'/>
-<snippet id='search-bar-import'/>
+The `SearchBar` module represents a UI component similar to `UISearchBar` in iOS and `	android.widget.SearchView` for Android, both of which allow you to to create a simple filter for the the content in the app. 
+This component provides `hint` and `text` properties and `submit` and `clear` events.
