+++ app/ns-ui-widgets-category/search-bar/search-bar-page.js
@@ -1,13 +1,12 @@
 const ListViewLinksModel = require("../../links-view-model");
 const link = require("../../link");
 const navigationLinks = [
-    new link("Basics", "ns-ui-widgets-category/search-bar/basics/basics-page"),
-    new link("Events", "ns-ui-widgets-category/search-bar/events/events-page"),
+    new link("Usage", "ns-ui-widgets-category/search-bar/usage/usage-page"),
+    new link("Styling", "ns-ui-widgets-category/search-bar/styling/styling-page"),
     new link("Code Behind", "ns-ui-widgets-category/search-bar/code-behind/code-behind-page")
 ];
 const navigationLinksTsc = [
-    new link("Basics", "ns-ui-widgets-category/search-bar/basics/basics-ts-page"),
-    new link("Events", "ns-ui-widgets-category/search-bar/events/events-ts-page"),
+    new link("Usage", "ns-ui-widgets-category/search-bar/usage/usage-ts-page"),
     new link("Code Behind", "ns-ui-widgets-category/search-bar/code-behind/code-behind-ts-page")
 ];
 function onNavigatingTo(args) {
