+++ app/ns-ui-widgets-category/search-bar/styling/article.md
@@ -1,0 +1,1 @@
+<snippet id='search-bar-styling-js'/>
