+++ app/ns-ui-widgets-category/search-bar/usage/article.md
@@ -1,0 +1,5 @@
+Using `SearchBar` component with `hint` and `text` with Angular binding. The user input can be handled by using the `submit`, `textChange` and `clear` events.
+
+<snippet id='search-bar-basics-xml'/>
+<snippet id='search-bar-basics-code'/>
+<snippet id='search-bar-basics-code-ts'/>
