+++ app/ns-ui-widgets-category/search-bar/usage/usage-ts-page.xml
@@ -1,0 +1,8 @@
+<Page xmlns="http://www.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo" loaded="onPageLoaded">
+    <Page.actionBar>
+      <ActionBar title="Getting Started"/>
+    </Page.actionBar>
+    <StackLayout>
+        <SearchBar id="searchBar" hint="Enter search term here ..." text="{{sbText}}" clear="onClear" submit="onSubmit" />
+    </StackLayout>
+</Page>
